// clients/js-kit/src/generated/types/plugin.ts
export type Plugin =
  | { __kind: 'Royalties'; fields: readonly [BaseRoyalties] }
  | { __kind: 'FreezeDelegate'; fields: readonly [FreezeDelegate] }
  | { __kind: 'BurnDelegate'; fields: readonly [BurnDelegate] }
  | { __kind: 'TransferDelegate'; fields: readonly [TransferDelegate] }
  | { __kind: 'UpdateDelegate'; fields: readonly [UpdateDelegate] }
  | {
      __kind: 'PermanentFreezeDelegate';
      fields: readonly [PermanentFreezeDelegate];
    }
  | { __kind: 'Attributes'; fields: readonly [Attributes] }
  | {
      __kind: 'PermanentTransferDelegate';
      fields: readonly [PermanentTransferDelegate];
    }
  | {
      __kind: 'PermanentBurnDelegate';
      fields: readonly [PermanentBurnDelegate];
    }
  | { __kind: 'Edition'; fields: readonly [Edition] }
  | { __kind: 'MasterEdition'; fields: readonly [BaseMasterEdition] }
  | { __kind: 'AddBlocker'; fields: readonly [AddBlocker] }
  | { __kind: 'ImmutableMetadata'; fields: readonly [ImmutableMetadata] }
  | { __kind: 'VerifiedCreators'; fields: readonly [VerifiedCreators] }
  | { __kind: 'Autograph'; fields: readonly [Autograph] }
  | { __kind: 'BubblegumV2'; fields: readonly [BubblegumV2] }
  | { __kind: 'FreezeExecute'; fields: readonly [FreezeExecute] }
  | {
      __kind: 'PermanentFreezeExecute';
      fields: readonly [PermanentFreezeExecute];
    };