// clients/js-kit/src/generated/types/externalPluginAdapterSchema.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from '@solana/kit';

export enum ExternalPluginAdapterSchema {
  Binary,
  Json,
  MsgPack,
}

export type ExternalPluginAdapterSchemaArgs = ExternalPluginAdapterSchema;

export function getExternalPluginAdapterSchemaEncoder(): FixedSizeEncoder<ExternalPluginAdapterSchemaArgs> {
  return getEnumEncoder(ExternalPluginAdapterSchema);
}

export function getExternalPluginAdapterSchemaDecoder(): FixedSizeDecoder<ExternalPluginAdapterSchema> {
  return getEnumDecoder(ExternalPluginAdapterSchema);
}

export function getExternalPluginAdapterSchemaCodec(): FixedSizeCodec<
  ExternalPluginAdapterSchemaArgs,
  ExternalPluginAdapterSchema
> {
  return combineCodec(
    getExternalPluginAdapterSchemaEncoder(),
    getExternalPluginAdapterSchemaDecoder()
  );
}