// clients/js-kit/src/generated/types/pluginAuthorityPair.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getOptionDecoder,
  getOptionEncoder,
  getStructDecoder,
  getStructEncoder,
  type Codec,
  type Decoder,
  type Encoder,
  type Option,
  type OptionOrNullable,
} from '@solana/kit';
import {
  getBasePluginAuthorityDecoder,
  getBasePluginAuthorityEncoder,
  getPluginDecoder,
  getPluginEncoder,
  type BasePluginAuthority,
  type BasePluginAuthorityArgs,
  type Plugin,
  type PluginArgs,
} from '.';

export type PluginAuthorityPair = {
  plugin: Plugin;
  authority: Option<BasePluginAuthority>;
};

export type PluginAuthorityPairArgs = {
  plugin: PluginArgs;
  authority: OptionOrNullable<BasePluginAuthorityArgs>;
};

export function getPluginAuthorityPairEncoder(): Encoder<PluginAuthorityPairArgs> {
  return getStructEncoder([
    ['plugin', getPluginEncoder()],
    ['authority', getOptionEncoder(getBasePluginAuthorityEncoder())],
  ]);
}

export function getPluginAuthorityPairDecoder(): Decoder<PluginAuthorityPair> {
  return getStructDecoder([
    ['plugin', getPluginDecoder()],
    ['authority', getOptionDecoder(getBasePluginAuthorityDecoder())],
  ]);
}

export function getPluginAuthorityPairCodec(): Codec<
  PluginAuthorityPairArgs,
  PluginAuthorityPair
> {
  return combineCodec(
    getPluginAuthorityPairEncoder(),
    getPluginAuthorityPairDecoder()
  );
}