// clients/js-kit/src/plugins/types.ts
export type FreezeDelegatePlugin = BasePlugin & FreezeDelegate;
export type BurnDelegatePlugin = BasePlugin & BurnDelegate;
export type TransferDelegatePlugin = BasePlugin & TransferDelegate;
export type UpdateDelegatePlugin = BasePlugin & UpdateDelegate;
export type PermanentFreezeDelegatePlugin = BasePlugin & PermanentFreezeDelegate;
export type AttributesPlugin = BasePlugin & Attributes;
export type PermanentTransferDelegatePlugin = BasePlugin & PermanentTransferDelegate;
export type PermanentBurnDelegatePlugin = BasePlugin & PermanentBurnDelegate;
export type EditionPlugin = BasePlugin & Edition;
export type MasterEditionPlugin = BasePlugin & MasterEdition;
export type AddBlockerPlugin = BasePlugin & AddBlocker;
export type ImmutableMetadataPlugin = BasePlugin & ImmutableMetadata;
export type VerifiedCreatorsPlugin = BasePlugin & VerifiedCreators;
export type AutographPlugin = BasePlugin & Autograph;
export type BubblegumV2Plugin = BasePlugin & BubblegumV2;
export type FreezeExecutePlugin = BasePlugin & FreezeExecute;
export type PermanentFreezeExecutePlugin = BasePlugin & PermanentFreezeExecute;

export type CommonPluginsList = {
  attributes?: AttributesPlugin;
  royalties?: RoyaltiesPlugin;
  updateDelegate?: UpdateDelegatePlugin;
  permanentFreezeDelegate?: PermanentFreezeDelegatePlugin;
  permanentTransferDelegate?: PermanentTransferDelegatePlugin;
  permanentBurnDelegate?: PermanentBurnDelegatePlugin;
  addBlocker?: AddBlockerPlugin;
  immutableMetadata?: ImmutableMetadataPlugin;
  autograph?: AutographPlugin;
  verifiedCreators?: VerifiedCreatorsPlugin;
  freezeExecute?: FreezeExecutePlugin;
  permanentFreezeExecute?: PermanentFreezeExecutePlugin;
};

export type AssetPluginsList = {
  freezeDelegate?: FreezeDelegatePlugin;
  burnDelegate?: BurnDelegatePlugin;
  transferDelegate?: TransferDelegatePlugin;
  edition?: EditionPlugin;
} & CommonPluginsList;

export type CollectionPluginsList = {
  masterEdition?: MasterEditionPlugin;
  bubblegumV2?: BubblegumV2Plugin;
} & CommonPluginsList;

export type PluginsList = AssetPluginsList & CollectionPluginsList;

export type PluginAuthorityPairHelperArgs = CreatePluginArgs & {
  authority?: BasePluginAuthority | PluginAuthority;
};

export type CreatePluginArgs =
  | {
      type: 'Royalties';
      data: BaseRoyaltiesArgs;
    }
  | {
      type: 'FreezeDelegate';
      data: FreezeDelegateArgs;
    }
  | {
      type: 'BurnDelegate';
    }
  | {
      type: 'TransferDelegate';
    }
  | {
      type: 'UpdateDelegate';
      data?: UpdateDelegateArgs;
    }
  | {
      type: 'Attributes';
      data: AttributesArgs;
    }
  | {
      type: 'PermanentFreezeDelegate';
      data: PermanentFreezeDelegateArgs;
    }
  | {
      type: 'PermanentTransferDelegate';
    }
  | {
      type: 'PermanentBurnDelegate';
    }
  | {
      type: 'Edition';
      data: EditionArgs;
    }
  | {
      type: 'MasterEdition';
      data: BaseMasterEditionArgs;
    }
  | {
      type: 'ImmutableMetadata';
    }
  | {
      type: 'AddBlocker';
    }
  | {
      type: 'BubblegumV2';
    }
  | {
      type: 'FreezeExecute';
      data: FreezeExecuteArgs;
    }
  | {
      type: 'PermanentFreezeExecute';
      data: PermanentFreezeExecuteArgs;
    }
  | {
      type: 'Autograph';
      data: AutographArgs;
    }
  | {
      type: 'VerifiedCreators';
      data: VerifiedCreatorsArgs;
    };

export type AuthorityArgsV2 = {
  authority?: PluginAuthority;
};

export type CreateOnlyPluginArgsV2 =
  | ({
      type: 'PermanentFreezeDelegate';
    } & PermanentFreezeDelegateArgs)
  | {
      type: 'PermanentTransferDelegate';
    }
  | {
      type: 'PermanentBurnDelegate';
    }
  | ({
      type: 'Edition';
    } & EditionArgs)
  | {
      type: 'BubblegumV2';
    }
  | ({
      type: 'PermanentFreezeExecute';
    } & PermanentFreezeExecuteArgs);

export type OwnerManagedPluginArgsV2 =
  | ({
      type: 'FreezeDelegate';
    } & FreezeDelegateArgs)
  | {
      type: 'BurnDelegate';
    }
  | {
      type: 'TransferDelegate';
    }
  | ({
      type: 'Autograph';
    } & AutographArgs)
  | ({
      type: 'FreezeExecute';
    } & FreezeExecuteArgs);

export type AuthorityManagedPluginArgsV2 =
  | ({
      type: 'Royalties';
    } & RoyaltiesArgs)
  | ({
      type: 'UpdateDelegate';
    } & UpdateDelegateArgs)
  | ({
      type: 'Attributes';
    } & AttributesArgs)
  | ({
      type: 'MasterEdition';
    } & MasterEditionArgs)
  | {
      type: 'ImmutableMetadata';
    }
  | {
      type: 'AddBlocker';
    }
  | ({
      type: 'VerifiedCreators';
    } & VerifiedCreatorsArgs);

export type AssetAddablePluginArgsV2 =
  | OwnerManagedPluginArgsV2
  | AuthorityManagedPluginArgsV2;
export type AssetAllPluginArgsV2 =
  | AssetAddablePluginArgsV2
  | CreateOnlyPluginArgsV2;
export type AssetPluginAuthorityPairArgsV2 = AssetAllPluginArgsV2 &
  AuthorityArgsV2;